# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Octo Protocol API
  version: 1.0.0
  description: >
    OpenAPI contract for the /v1 Octo Protocol routes.


    Every `amount_stroops` field below is a JSON `integer` and stays that way in this version —
    that is unchanged by issue #215. Storage now also carries an arbitrary-precision
    `amount_base_units NUMERIC(78,0)` column internally (needed so a future, non-Stellar chain
    whose amounts don't fit 64 bits can be represented exactly), but nothing about that is exposed
    on these schemas yet. The public cutover to a string-typed amount field is a separate,
    versioned breaking change tracked in issue #227 — do not assume `amount_stroops` stays numeric
    across a future major version.
servers:
  - url: /
paths:
...
  /v1/wallets/{id}/addresses:
    post:
      summary: Generate a dedicated customer address
      operationId: createAddress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                customer_ref:
                  type: string
                metadata:
                  type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateAddressResponse'
    get:
      summary: List addresses
      operationId: listAddresses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAddressesResponse'
...
components:
  schemas:
...
    AddressView:
      type: object
      description: >-
        Response shape depends on chain_kind (see docs/deposit-model.md). Stellar responses carry
        muxed_address/base_address/memo_id; EVM responses carry address instead and OMIT
        memo_id entirely (not null) — EVM has nowhere for a memo to go.
      required:
        - id
        - chain_kind
        - metadata
        - received_stroops
      properties:
        id:
          type: string
          format: uuid
        customer_ref:
          type: string
        chain_kind:
          type: string
          enum: [stellar, evm]
        muxed_address:
          type: string
          description: Stellar only.
        base_address:
          type: string
          description: Stellar only.
        memo_id:
          type: integer
          description: Stellar only. Never present on an EVM response.
        address:
          type: string
          description: EVM only. EIP-55 checksummed deposit address.
        metadata:
          type: object
        received_stroops:
          type: integer
    CreateAddressResponse:
      type: object
      required:
        - statusCode
        - message
        - data
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          $ref: '#/components/schemas/AddressView'
    ListAddressesResponse:
      type: object
      required:
        - statusCode
        - message
        - data
      properties:
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/AddressView'
...